+++ napari/_tests/test_with_screenshot.py
@@ -8,8 +8,8 @@ from napari._tests.utils import (
     skip_on_mac_ci,
     skip_on_win_ci,
 )
+from napari.utils._proxies import ReadOnlyWrapper
 from napari.utils.interactions import (
-    ReadOnlyWrapper,
     mouse_move_callbacks,
     mouse_press_callbacks,
     mouse_release_callbacks,
