+++ napari/components/_tests/test_viewer_mouse_bindings.py
@@ -4,7 +4,8 @@ import numpy as np
 import pytest
 
 from napari.components import ViewerModel
-from napari.utils.interactions import ReadOnlyWrapper, mouse_wheel_callbacks
+from napari.utils._proxies import ReadOnlyWrapper
+from napari.utils.interactions import mouse_wheel_callbacks
 
 
 @pytest.fixture
