+++ napari/utils/_tests/test_interactions.py
@@ -1,28 +1,6 @@
 import pytest
 
-from ..interactions import ReadOnlyWrapper, Shortcut
-
-
-def test_ReadOnlyWrapper_setitem():
-    """test that ReadOnlyWrapper prevents setting items"""
-    d = {'hi': 3}
-    d_read_only = ReadOnlyWrapper(d)
-
-    with pytest.raises(TypeError):
-        d_read_only['hi'] = 5
-
-
-def test_ReadOnlyWrapper_setattr():
-    """test that ReadOnlyWrapper prevents setting attributes"""
-
-    class TestClass:
-        x = 3
-
-    tc = TestClass()
-    tc_read_only = ReadOnlyWrapper(tc)
-
-    with pytest.raises(TypeError):
-        tc_read_only.x = 5
+from ..interactions import Shortcut
 
 
 @pytest.mark.parametrize(
