+++ napari/utils/_tests/test_proxies.py
@@ -1,0 +1,25 @@
+import pytest
+
+from napari.utils._proxies import ReadOnlyWrapper
+
+
+def test_ReadOnlyWrapper_setitem():
+    """test that ReadOnlyWrapper prevents setting items"""
+    d = {'hi': 3}
+    d_read_only = ReadOnlyWrapper(d)
+
+    with pytest.raises(TypeError):
+        d_read_only['hi'] = 5
+
+
+def test_ReadOnlyWrapper_setattr():
+    """test that ReadOnlyWrapper prevents setting attributes"""
+
+    class TestClass:
+        x = 3
+
+    tc = TestClass()
+    tc_read_only = ReadOnlyWrapper(tc)
+
+    with pytest.raises(TypeError):
+        tc_read_only.x = 5
